# rally2-matching-system/api/matching-system-swagger.yml
swagger: "2.0"

info:
  title: The Maryland Test Facility Matching System Interface
  description: Application Programming Interface for receiving matching system requests from the MdTF Backend.
  version: "1.0.3"
  contact:
    name: John Howard
    url: https://mdtf.org
    email: john@mdtf.org
  # the license of the swagger documentation
  license:
    name: Copyright (c) 2018, The Maryland Test Facility
# the domain of the service
host: 172.17.0.2:8080
# array of all schemes that your API supports
schemes:
  - http
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json

################################################################################
#                                   Tags                                       #
################################################################################

tags:
  - name: Biometric Operations
  - name: Algorithm Information

################################################################################
#                                   Definitions                                #
################################################################################

definitions:

  ############
  ## REQUESTS
  ############
  Image:
    type: object
    description: |
      Data transfer object for an image.
    properties:
      ImageData:
        type: string
        description: The captured image data in PNG format, encoded as a base64 string. The data string shall not exceed 1MB.
        example: iVBORw0KGgoAAAANSUhEUgAAAAEAAAABCAIAAACQd1PeAAAAEElEQVR4nGJiYGAABAAA//8ADAADcZGLFwAAAABJRU5ErkJggg==
        x-nullable: false
    required:
      - ImageData

  Template:
    type: object
    description: |
      Data transfer object for a template.
    properties:
      Template:
        description: The template data, encoded as a base64 string.  The data string shall not exceed 1 MB.
        type: string
        example: dGhpcyBzZW50ZW5jZSBpcyBhbiBleGFtcGxlIHRlbXBsYXRlLi4K
        x-nullable: false
    required:
      - Template

  Comparison:
    type: object
    description: |
      A similarity score for a single (1:1) template comparison operation.
    properties:
      Score:
        type: number
        format: float
        description: An un-normalized similarity score, as produced by the algorithm
        example: 8734
        x-nullable: false
      NormalizedScore:
        type: number
        format: float
        description: Similarity score between 0 and 1, with 1 being the highest score the algorithm can produce
        example: 0.8734
        x-nullable: false
    required:
      - Score
      - NormalizedScore

  CompareListRequest:
    description: A single template object and a list of templates that it will be compared to.
    type: object
    properties:
      SingleTemplate:
        $ref: "#/definitions/Template"
      TemplateList:
        type: array
        items:
          $ref: "#/definitions/Template"
        x-nullable: false
    required:
      - SingleTemplate
      - TemplateList

  #############
  ## RESPONSES
  #############

  Info:
    type: object
    description: Basic information describing the algorithm.
    properties:
      AlgorithmName:
        type: string
        description: Name of algorithm
        example: AlwaysTrue
      AlgorithmVersion:
        type: string
        description: Algorithm version identifier
        example: "1.0.1"
      AlgorithmType:
        type: string
        enum: [Face, Finger, Iris]
        description: A string enum describing the type of biometric images the algorithm is meant to process
      CompanyName:
        type: string
        description: Name of the Company which produces the algorithm
        example: MdTF
      TechnicalContactEmail:
        type: string
        description: The email address of an engineer or other technical resource to contact in the event of an error running your service. This field may be left blank if desired.
        example: john@mdtf.org
      RecommendedCPUs:
        type: number
        format: integer
        description: The recommended allocation of CPUs for the deployed docker container.
        example: 4
      RecommendedMem:
        type: number
        format: integer
        description: The recommended allocation of memory (MB) for the deployed docker container.
        example: 2048
      Thresholds:
        description: A map of thresholds to their expected values. Note that thresholds values cannot exceed a length of 10 characters.
          There are 5 required thresholds, and you may specify 5 additional thresholds.
        maxProperties: 10
        type: object
        properties:
          "1:500":
            type: string
            maxLength: 10
          "1:1e3":
            type: string
            maxLength: 10
          "1:1e4":
            type: string
            maxLength: 10
          "1:1e5":
            type: string
            maxLength: 10
          "1:1e6":
            type: string
            maxLength: 10
        required:
          - "1:500"
          - "1:1e3"
          - "1:1e4"
          - "1:1e5"
          - "1:1e6"
        additionalProperties:
          type: string
          maxLength: 10
        example:
          "1:500": "0.99"
          "1:1e3": "0.97"
          "1:1e4": "0.96"
          "1:1e5": "0.86"
          "1:1e6": "0.753"
    required:
      - AlgorithmName
      - AlgorithmVersion
      - AlgorithmType
      - CompanyName
      - TechnicalContactEmail
      - RecommendedCPUs
      - RecommendedMem
      - Thresholds

  RequestError:
    type: string
    description: |
      Relevant and concise diagnostic information in the event of a client side error (e.g. malformed requests, or invalid image encoding).
    example: Unable to decode image data as a PNG.

  ServerError:
    type: string
    description: |
      Relevant and concise diagnostic information in the event of a server side error (e.g. invalid license, or failure to generate a template).
    example: The internal license has expired.
################################################################################
#                                   Parameters                                 #
################################################################################

parameters:
  image:
    name: image
    in: body
    description: |
      The biometric image that is being submitted for feature extraction.
    required: true
    schema:
      $ref: "#/definitions/Image"

  compare_list_request:
    name: compare_list_request
    in: body
    description: A single template object and a list of templates that it will be compared to.
    required: true
    schema:
      $ref: "#/definitions/CompareListRequest"

################################################################################
#                                   Paths                                      #
################################################################################

paths:
  /v1/create-template:
    post:
      summary: Generate a template from the provided biometric image
      description: |
        This endpoint accepts a base64 encoded PNG and attempts to perform a 'feature extraction' operation producing a single template.
      tags:
        - Biometric Operations
      operationId: create_template
      parameters:
        - $ref: "#/parameters/image"
      responses:
        # Response code
        200:
          description: Successful Response
          schema:
            $ref: "#/definitions/Template"
        400:
          description: Bad Request
          schema:
            $ref: "#/definitions/RequestError"
        500:
          description: Internal Server Error
          schema:
            $ref: "#/definitions/ServerError"

  /v1/compare-list:
    post:
      summary: Compare a single template to a list of templates
      description: |
        This endpoint accepts a template and a list of templates. It compares the single template to every template in the provided list. The result is a list of Comparison objects that holds a similarity score for each comparison. <br><br> The returned comparison list MUST contain the same number of elements AND be in the same order as the provided list of templates.
      tags:
        - Biometric Operations
      operationId: compare_list
      parameters:
        - $ref: "#/parameters/compare_list_request"
      responses:
        # Response code
        200:
          description: Successful Response
          schema:
            type: array
            description: An array of comparison results. This list MUST contain the same number of elements AND be in the same order as the provided list of templates.
            items:
              $ref: "#/definitions/Comparison"
        400:
          description: Bad Request
          schema:
            $ref: "#/definitions/RequestError"
        500:
          description: Internal Server Error
          schema:
            $ref: "#/definitions/ServerError"

  /v1/info:
    get:
      summary: Returns basic information for the algorithm.
      description: |
        This endpoint returns some basic information about the algorithm. The response should be stateless, with no fixed values.
      tags:
        - Algorithm Information
      operationId: info
      responses:
        # Response code
        200:
          description: Successful Response
          schema:
            $ref: '#/definitions/Info'
        500:
          description: Internal Server Error
          schema:
            $ref: "#/definitions/ServerError"